Moved AdminClientProvider and default implementation in operator common package
