+++ operator-common/src/main/java/io/strimzi/operator/common/AdminClientProvider.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.operator.cluster.operator.resource;
+package io.strimzi.operator.common;
 
 import io.fabric8.kubernetes.api.model.Secret;
 import org.apache.kafka.clients.admin.AdminClient;
